+++ rust/component/src/parameters.rs
@@ -28,10 +28,8 @@ use std::{
     string::ToString,
 };
 
-#[cfg(any(test, feature = "test-utils"))]
-pub mod test_utils;
-
-pub mod utils;
+mod utils;
+pub use utils::*;
 
 #[cfg(test)]
 mod tests;
